Abstract Container
Containers (Inventories) now exist, and work (control) without any UI attachment. They are set up in a way that should allow for saving inventory contents easier. 

This needs some sort of Slot monobehaviour which can read the contents of the list of SlotData stored on the container, and render it appropriately. Shouldn't be too difficult.,